Renaming and Hierarchy Organization
Renamed all objects in greybox folder so it's easier for us to find all of the objects; if need be, I can write out a key or doc for finding objects, because eventually the project file is going to get very populated and finding things will be hard. I also reorganized the hierarchy to reflect these changes. I've based the naming and hierarchy changes off of the two different paths that users will be able to take. I named them Paths A & B. Path A is the one where users will go left to start and Path B is the one where users will go down; these directions are referenced from the map of the environment that was provided