Imported Assets, Added Textures
I added textures to the mountains in the background and flattened them a bit so they could be more level with one another. I also added a quick ocean to make the water less boring. To make the environment a little more engaging, I also added an audio source and beach sound effect to the water and made it very quiet so it's a constant ambience. More changes to come!